+++ src/main/java/com/example/eccomerce/model/dtos/response/ResponseItemCartDto.java
@@ -4,11 +4,13 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
+
 public record ResponseItemCartDto(
         String id,
         String cartId,
         String productId,
         int quantity,
-        int subTotal
+        BigDecimal subTotal
 ) {
 }
